파일명 변경 및 불필요 파일 삭제
